+++ src/cryptography/hazmat/backends/openssl/utils.py
@@ -64,6 +64,6 @@ def _warn_sign_verify_deprecated():
     warnings.warn(
         "signer and verifier have been deprecated. Please use sign "
         "and verify instead.",
-        utils.PersistentlyDeprecated,
+        utils.PersistentlyDeprecated2017,
         stacklevel=3
     )